chore(tests): logically organize tests
